refactor/extract HomeController.add() to TaskController
