+++ app/presentation/controllers/TaskController.scala
@@ -1,0 +1,29 @@
+/**
+  * to do sample project
+  */
+
+package presentation.controllers
+
+import ixias.play.api.mvc.JsonHelper
+import play.api.libs.json.Json
+import play.api.mvc._
+import usecase.task.AddTaskUseCase
+
+import javax.inject._
+import scala.concurrent.Future
+
+@Singleton
+class TaskController @Inject() (
+  val controllerComponents: ControllerComponents,
+  addTaskUseCase:           AddTaskUseCase
+) extends BaseController {
+  implicit val ec: scala.concurrent.ExecutionContext = controllerComponents.executionContext
+
+  def add() = Action.async { implicit request =>
+    JsonHelper.bindFromRequest[AddTaskRequest].fold(
+      formWithErrors => Future(formWithErrors),
+      addTaskRequest =>
+        addTaskUseCase.execute(addTaskRequest).map(task => Ok(Json.toJson(task)))
+    )
+  }
+}
